fix(nvim): use a workaround for `Lazy sync`
When installing `bracey.vim` and `markdown-preview.nvim`, the building
process creates extra files that are considered new by git. That way
the git pull, which happens on `Lazy sync` fails to work. This workaround
changes the building process for each of the plugins to disable the
creation of new files.