fix: finally fix the keymap bug in neovim and vim
The previous version had `"_d` before pasting, which for some reason works
different when used on selected text at the end of the line, removing
extra character and ruining my experience. All I wanted was to not
modify the register when pasting in visual mode, which is the default
behavior for `p`. As I accidentally discovered, `P` does exactly what I
need. Great! Makes no sense but I'm happy!